requirements.txt
agate==1.7.1
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.9
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
bleach==6.0.0
Brotli==1.0.9
cachetools==5.3.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.26.3
comm==0.1.3
constantly==15.1.0
contextlib2==21.6.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
daphne==4.0.0
dbfread==2.0.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.2.1
django-admin-log==0.2
django-celery==3.1.17
django-celery-results==2.5.0
django-channels==0.7.0
django-clearcache==1.2.1
django-cors-headers==3.14.0
django-db-logger==0.1.12
django-post-request-task==0.5
django-redis==5.2.0
djangomix==1.1.3
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
ffmpeg==1.4
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
geoip2==4.6.0
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
grad-cam==1.4.6
greenlet==2.0.2
grpcio==1.51.3
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.23.0
ipython==8.12.2
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.18.2
Jinja2==2.11.3
joblib==1.2.0
json5==0.9.11
jsonschema==4.17.3
jupyter-server==1.24.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab==3.5.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.10.3
kiwisolver==1.4.4
kombu==5.2.4
lazy_loader==0.2
leather==0.3.4
librosa==0.10.0.post2
lit==16.0.3
llvmlite==0.40.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
maxminddb==2.2.0
mistune==0.8.4
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mutagen==1.46.0
mysqlclient==2.1.1
nbclassic==1.0.0
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile==0.46
opencv-python==4.7.0.72
openpyxl==3.1.2
opensoundscape==0.9.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parsedatetime==2.6
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.0
pluggy==1.0.0
pooch==1.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.17
Pygments==2.15.1
pylzma==0.5.0
pymemcache==4.0.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
ray==2.4.0
redis==4.5.4
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
schema==0.7.5
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
Send2Trash==1.8.2
sentry-sdk==1.22.2
service-identity==21.1.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
SQLAlchemy==1.4.0
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.11.1
terminado==0.17.1
testpath==0.6.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.4.12
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
ttach==0.0.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.15
user-agents==2.2.0
vine==5.0.0
virtualenv==20.21.0
wandb==0.13.11
wcwidth==0.2.6
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.5.1
websockets==11.0.3
webvtt-py==0.4.6
wrapt==1.15.0
xlrd==2.0.1
yarl==1.9.2
yt-dlp==2023.3.4
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
